Implemented In-Memory Caching Mechanism using Redis and performed some Refactoring

+++ FirstApiSandbox/Service/BookService.cs
@@ -6,19 +6,21 @@ using FirstApiSandbox.Model;
 using FirstApiSandbox.Database;
 using FirstApiSandbox.Validation;
 using FluentValidation.Results;
+using ServiceStack.Redis;
 
 namespace FirstApiSandbox.Service
 {
     public class BookService : IService
     {
         private IBookDatabase bookData;
+        RedisManagerPool redisManager = new RedisManagerPool("localhost:6379");
+        IRedisClient redisClient;
+
         public BookService(IBookDatabase bookDatabase)
         {
             bookData = bookDatabase;
+            redisClient = redisManager.GetClient();
         }
-
-        public const string Valid = "Valid book";
-        //BookData bookData = new BookData();
 
         public Response GetBooksfromService()
         {
@@ -29,35 +31,67 @@ namespace FirstApiSandbox.Service
 
         public Response GetBookFromServiceByName(string name)
         {
-            var returnedBook = bookData.GetBooksFromDatabaseByName(name);
-            Response response = new Response(returnedBook, Errors.NotFound);
+            if (BookValidation.IfQueriedBookNameExists(name))
+            {
+                if (redisClient.Get<Book>(name) != null)
+                {
+                    //Book found in Cache
+                    var returnedBook = redisClient.Get<Book>(name);
+                    Response response = new Response(returnedBook);
+                    return response;
+                }
+                else
+                {
+                    //Book found in Database
+                    var returnedBook = bookData.GetBooksFromDatabaseByName(name);
+                    redisClient.Set(returnedBook.Name, returnedBook, TimeSpan.FromMinutes(1));
+                    Response response = new Response(returnedBook);
+                    return response;
+                }
+            }
+            else
+            {
+                Response response = new Response(Errors.NotFound);
+                return response;
+            }
 
-            return response;
+
+            #region  Safe Code
+            //if (BookValidation.IfQueriedBookNameExists(name))
+            //{
+            //    var returnedBook = bookData.GetBooksFromDatabaseByName(name);
+            //    Response response = new Response(returnedBook);
+
+            //    return response;
+            //}
+            //else
+            //{
+            //    Response response = new Response(Errors.NotFound);
+            //    return response;
+            //}     
+            #endregion
         }
 
         public Response AddBookUsingService(Book newBook)
         {
             Response response;
 
-            #region Experimenting Fluent Validation *Commented code*
             BookValidation bookValidator = new BookValidation();
             ValidationResult result = bookValidator.Validate(newBook);
-            //ruleSet: "MustNotBeEmpty"
 
             if (result.IsValid)
             {
                 bookData.AddBookInDatabase(newBook);
                 var returnedBook = bookData.GetBooksFromDatabaseByName(newBook.Name);
-                response = new Response(returnedBook, null);
+                response = new Response(returnedBook);
                 return response;
             }
             else
             {
                 IList<ValidationFailure> validationFailureMessages = result.Errors;
-                response = new Response(validationFailureMessages);
+                response = new Response(validationFailureMessages[0].ErrorMessage);
                 return response;
             }
-            #endregion
 
             #region Experimenting Without Fluent Validation *Commented code*
             //string validationCode = BookValidation.IsBookValid(newBook);
@@ -88,45 +122,40 @@ namespace FirstApiSandbox.Service
                 if (result.IsValid)
                 {
                     bookData.UpdateBookInDatabase(name, newBook);
+                    if (redisClient.Get<Book>(name) != null)
+                    {
+                        redisClient.DeleteById<Book>(name);
+                        redisClient.Set(newBook.Name, newBook, TimeSpan.FromMinutes(1));
+                    }
                     Book returnedBook = bookData.GetBooksFromDatabaseByName(newBook.Name);
-                    response = new Response(returnedBook, null);
+                    response = new Response(returnedBook);
                     return response;
                 }
                 else
                 {
                     IList<ValidationFailure> validationFailureMessages = result.Errors;
-                    response = new Response(validationFailureMessages);
+                    response = new Response(validationFailureMessages[0].ErrorMessage);
                     return response;
                 }
             }
             else
-            {
-                response = new Response(null, Errors.NotFound);
-                return response;
-            }
-
-            
-            #region Experimenting Without Fluent Validation *Commented code*
-            //string validationCode = BookValidation.IsBookValid(newBook);
-            //if (validationCode.Equals(Valid))
-            //{
-            //    bookData.UpdateBookInDatabase(name, newBook);
-            //    Book returnedBook = bookData.GetBooksFromDatabaseByName(newBook.Name);
-            //    response = new Response(returnedBook, null);
-            //    return response;
-            //}
-            //else
-            //{
-            //    response = new Response(null, validationCode);
-            //    return response;
-            //}
-            #endregion
+                return new Response(Errors.NotFound); ;
         }
 
         public Response DeleteBookUsingService(string name)
         {
-            bookData.DeleteBookFromDatabase(name);
-            return new Response(null, Errors.DeletionSuccessful);
+            if (BookValidation.IfQueriedBookNameExists(name))
+            {
+                bookData.DeleteBookFromDatabase(name);
+                if (redisClient.Get<Book>(name) != null)
+                {
+                    redisClient.DeleteById<Book>(name);
+                }
+                return new Response(Errors.DeletionSuccessful);
+            }
+            else
+                return new Response(Errors.NotFound);
+
         }
 
         public Response GetBookByGenreUsingService(string genreName)
